+++ example_project/HelloScript.kt
@@ -75,6 +75,8 @@ import net.multigesture.kanama.api.Texture2D
 import net.multigesture.kanama.api.Time
 import net.multigesture.kanama.api.Timer
 import net.multigesture.kanama.api.Tween
+import net.multigesture.kanama.api.createTween
+import net.multigesture.kanama.api.getProcessedTweens
 import net.multigesture.kanama.binding.runtime.ObjectCalls
 import net.multigesture.kanama.generated.HelloScriptNames
 import net.multigesture.kanama.generated.HelloScriptSignals